backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP Client
httpx==0.28.1

# Environment Variables
python-dotenv==1.0.1

# Gemini SDK
google-generativeai==0.8.3

# Utilities
python-multipart==0.0.18

# PowerPoint Generation
python-pptx==0.6.23

# Word Document Generation
python-docx==1.1.2
